change error code to MODID for unit tests

+++ web_frontend/cloudscheduler/csv2/cloud_views.py
@@ -59,7 +59,7 @@ CLOUD_KEYS = {
         'metadata_name':                        'ignore',
         'metadata_option':                      ['add', 'delete'],
         'ram_ctl':                              'integer',
-        'spot_price':                           'integer',
+        'spot_price':                           'float',
         'vm_keep_alive':                        'integer',
 
         'cores_slider':                         'ignore',
